Different parameter names from content-manager
The content-manager plugin routes use the parameter name id (:id) in routes for updating content from the admin portal. The user-permissions plugin uses _id  (:_id) as the parameter name in the routes. This causes a conflict as the content-manager calls the update function of the user controller with a different parameter.