Trimmed whitespace for AWS S3 credentials
If you input a name/key to a bucket with a trailing whitespace, it will
respect that trailing whitespace but will then fail to find the bucket.

To resolve this, added a util function to check if it's a string. If so
it will apply the `trim()` function to remove any unwanted whitespace,
else it will return `undefined`.