feat: add "create new" functonality to SelectMany
This commit introduces a "Create new" link to the SelectMany components
provided the following are true:

- The relation is of type "oneToMany" (that is, "Belongs to many" in the
  content type builder, or a bidirectional relation).
- The item has already been created and is being edited.

This enables a simpler workflow: previously, one would have to go to the
ListView of the related model, create a new entry, save it, go back to
the item we were editing, and associate the newly created entry to the
item.

Now, the user can click the "Create new" link, create the new item, and
it will be automatically associated to the source item. When the user
saves it, they will be redirected back to the EditView of the source
item.

This only works on "Belongs to many" relations, and not in "Has many",
because the association is made from the target model's creation view,
and not from the source model's edit view. Therefore, we need the
relation to be bidirectional.

Signed-off-by: Pablo Rodríguez Caballero <pablo@rodriguezcaballero.me>