Updating s3 provider to add an Cloudfront option
Updating provider configuration in order to be able to enable Cloudfront delivery for uploaded assets to s3 bucket. This configuration add a Yes / No dropdown to indicate if we wish to store the Cloudfront URL (which is the other configurable field) of the uploaded asset instead of the S3 bucket URL. This entails that the user has already created a Cloudfront distribution for his/her s3 bucket.