Translations: Updated the 'terms' translations for 'de'
"Terms and Conditions" are called "Allgemeine Geschäftsbedingungen" in
German (often abbreviated as "AGB") or "Nutzungsbedingungen".

In the context of the admin sign-up of 'strapi' I'd go for the latter.
(Research results attached to the pull request)

The previous version "Begriffe" is wrong in this context as it is the
literal translation of the word "term".

Signed-off-by: Lennart Stolz <lennartstolz@me.com>