Fix mongoose example to add necessary .exec() step
Mongoose's model.find() method returns a `Promise` that when resolved is a `model.Query` object. The `model.Query` must be `.exec()`'d before it can be transformed using `.toObject()`.

Without this change you get this error in the console: `error TypeError: result.toObject is not a function"`, and this error on the client side: `Failed to load resource: the server responded with a status of 400 (Bad Request)`.

In what I think is the spirit of the original example, to get the `fields = result.toObject()` of the returned result, I've also chained a `.lean()` suffix to the `.find()` call. Additionally I've changed the "fields" variable name to "results" as .find().lean() will return an array of js object results. 

To get a "fields" result You'd have to do something like this:
```js
fields = results[0];
``` 
... or better yet, to match the bookshelf example where a query is being made by id, change the example to `.findOne()` and `const fields = query.exec()`. Like this:
```js 
const query = strapi.query('restaurant').model.findOne({ _id: 1 }).lean();
const fields = await query.exec();
```

A secondary error that pops up when using the original example, `.find()`, after it's result is `.exec()`'d, returns an `Array` which will throws an error when you call `.toObject()` against it; same error as before. You could use `.findOne()` in the original example, which returns a single model entry, but then you're stuck without the data values that I think the spirit of the original example was trying to accomplish using `.toObject()`. Here's an example of what I mean: 

*using `.find().lean()` then `.exec()`*

```js 
[{
   _id:"5cf5c56e0962c88eed2506d9",
  issue_number": 2, // the long value as a number 
  createdAt:"2019-06-04T01:12:14.225Z" // the date value as a string
}]
```

*using `.find()`, then `.exec()`, then `.toObject()` 

```js
[{
   _id:"5c92b6d02fc8512d06d9e476",
  issue_number: Long {_bsontype: "Long", low_: 5, high_: 0} // a Long object
  createdAt: Mon Jun 08 2020 10:45:55 GMT-0700 (Pacific Daylight Time) {} // a Date Object
}]
```

Thanks for taking a look at this guys. Keep up the good work!