Update heroku.md
I have struggled with setting up PostgreSQL database on Heroku using this tutorial. It says that if you want to use PostgreSQL you should not use --quickstart option, however, since this is tutorial for NOT sqlite database, running npx with --quickstart then totally disables the developer to later change to postgesql as intended. Therefore, I would not suggest running the --quickstart option in this case.