Update README.md (#7190)
Added extra settings on the mailgun config based on the mailgun usage overview.
You can check the doc, and it specify the domain parameter.
https://www.npmjs.com/package/mailgun-js#usage-overview

You need to specify the domain, otherwise the email won't be sent, but the API return success anyway. 
In my scenario I have a mailgun account with multiple domains and without this param the email was not sent. Don't know if a mailgun with a single domain works.

Signed-off-by: Gil Fernandes <gil.fernandes@onepointltd.com>