Add customDomain option when uploading files to S3
Adding the `customDomain` config field to the `upload-aws-s3` provider. 

It allows you to specify a custom domain while saving the files' url. This is especially helpful when the media files are served through a CDN and you need the urls to point to that domain instead of the default bucket's one.