Config on router and handler config be nullable
Since you can pass certain methods in the only array it doesn't make sense to then have to pass in a config with every single handler config on methods that will not be used.  If I have a router that was setup to be only: ['find'] I shouldn't need to create a HandlerConfig for the other crud operations.  As well as on the handlerconfig on javascript version of Strapi none of those properties are required, especially the auth property where the default is enabled when you **DO NOT** pass a value, meaning we certainly should not have to always pass a value for it when using typescript.