Register logger middleware before errors middleware
In the examples and project templates the "errors" middleware which turn
thrown errors into HTTP responses is registered before the "logger"
middleware. This causes any errors thrown in controllers to pierce
through the logger middleware resulting in these requests not being
logged.

Eg. when a controller throws a ValidationError the resulting HTTP 400
request is not logged at all.

Change the order of middleware registration so that the logger is
'above' the errors middleware and has a chance to log *all* requests.