feat: add mapping property support for documentation plugin's open api generated spec
Dynamic zones are reflected in the openapi docs generated as a anyOf using the list of possible components.
For this to be more useful according the openapi spec we can use a discriminator to let the docs know what
property is responsible for the discrimination and the mapping between the property values and schemas.