fix(upload): empty folder list in root view
In the media library, the dropdown used to move assets was not displaying the children folders.
This was due to mixed values (`null` and `''`) while the check was using a strict comparison.

This PR fixes the issue by only using an empty string.
I chose to not use the `null` value since it's also used in react-select which does not accept it.