fix(admin): add `origin` to doc hooks
When cloning a document on
`/admin/content-manager/collection-types/:slug/clone/:origin`, we do not
set the document id to the origin document id since this is the creation
of a new document. But a document id is required to query relations and
without it, the field is kept empty.

This updates the `useDoc` hook to pass the `origin` string and use that
as an alternative document id.